requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-dotenv
anthropic
httpx
python-jose[cryptography]
bcrypt
gunicorn
pytest
python-dateutil